Add files via upload
It breaks Ftp in a few seconds and performs a bulk scan randomly or by defining a range. 

Use it in a simple way with few multiple selection commands.

modify the users and passwords within the solution.

Written entirely in c #

I am not responsible for any misuse or damage caused.